+++ Tests/UnitTests/TestHelpers/AsyncTestHelpers.swift
@@ -15,6 +15,8 @@ import Foundation
 
 import Nimble
 
+@testable import RevenueCat
+
 @available(iOS 13.0, macOS 10.15, tvOS 13.0, watchOS 6.2, *)
 internal extension AsyncSequence {
 
@@ -52,4 +54,40 @@ func waitUntilValue<Value>(
     }
 
     return value
+}
+
+/// Verifies that the given `async` condition becomes true after `timeout`,
+/// checking every `pollInterval`.
+// Note: a better approach would be using `XCTestExpectation` and `self.wait(for:timeout:)`
+// but it doesn't seem to play well with async-await.
+// Also `toEventually` (Quick nor Nimble) don't support `async`.
+// Fix-me: remove once we can use Quick v6.x:
+@available(iOS 13.0, macOS 10.15, tvOS 13.0, watchOS 6.2, *)
+func asyncWait(
+    until condition: @Sendable () async -> Bool,
+    timeout: DispatchTimeInterval,
+    pollInterval: DispatchTimeInterval,
+    description: String? = nil,
+    file: FileString = #fileID,
+    line: UInt = #line
+) async {
+    let start = Date()
+    var foundCorrectValue = false
+
+    func timedOut() -> Bool {
+        return DispatchTimeInterval(Date().timeIntervalSince(start)) > timeout
+    }
+
+    repeat {
+        foundCorrectValue = await condition()
+        if !foundCorrectValue {
+            try? await Task.sleep(nanoseconds: UInt64(pollInterval.nanoseconds))
+        }
+    } while !(foundCorrectValue || timedOut())
+
+    expect(
+        file: file,
+        line: line,
+        foundCorrectValue
+    ).to(beTrue(), description: description)
 }